Integrate VTPM2 with JSS
This change integrates vTPM2 support with existing JSS service. JSS becomes a wrapper to TPM
