--- components/vtpm2-server/run-server.sh
@@ -1,2 +1,0 @@
-#!/bin/sh
-FLASK_APP=/server.py python -m flask run --host=0.0.0.0
